I think this is how you bootstrap :|
Added bootstrapping which jumps to C as fast as possible.
This code was basically copied from http://wiki.osdev.org/Bare_Bones#Bootstrap_Assembly
and I may modify it later as needed. This does not fully initialize the
processor and that will happen later in C (so the FPU is still not initialized)