Create a memrev function
Created a memrev (memory reverse) function which is designed to reverse the memory block
in question. This replaces the previous strrev function internals. In the future, this can be
used to convert between big endian and little endian once an endianness test is implemented which
will be nice for saving binary chunks where endianness differs or emulating a specific endianness
of another system for an application or driver.